feat : remove image on course delete
