+++ app/dependencies/firebase.py
@@ -1,28 +1,11 @@
-import os.path
-
-from functools import lru_cache
 from fastapi import HTTPException
-
-from firebase_admin import initialize_app, credentials, auth
-from app.core.config import get_settings
-
+from firebase_admin import storage
 import jwt
 
 def check_token(token):
     return jwt.decode(token, "secret", algorithms=["HS256"])
 
-@lru_cache
-class FirebaseAuth():
-    def __init__(self):
-        super().__init__()
-        settings = get_settings()
-        if not settings.firebase_auth:
-            raise Exception("Firebase Credential Location not found in .env")
-        if not os.path.isfile(f"./{settings.firebase_auth}"):
-            raise Exception(f"Firebase Credential File not found in /app/{settings.firebase_auth}")
-        cred = credentials.Certificate(settings.firebase_auth)
-        initialize_app(cred)
-
+class FirebaseAuth:
     def verify_token(self, token):
         try:
             return check_token(token)["user_id"]
@@ -31,5 +14,15 @@ class FirebaseAuth():
     def develop_create_token(self, uid):
         return jwt.encode({"user_id": uid}, 'secret', algorithm="HS256")
 
-def get_auth():
+def get_auth() -> FirebaseAuth:
     return FirebaseAuth()
+
+# TODO : 만약에 Blob이 존재하지 않는다면..?
+class FirebaseStorage:
+    @classmethod
+    def remove_image_by_walk_id(self, walk_id):
+        prefix = "images/"
+        postfix = ".png"
+        bucket = storage.bucket()
+        blob = bucket.get_blob(prefix + walk_id + postfix)
+        blob.delete()
